pyproject.toml
[project]
name = "effi-mail"
version = "0.2.0"
description = "MCP server for Outlook email management with triage workflow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "pywin32>=306",
]

[project.scripts]
effi-mail = "effi_mail:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
